requirements.txt
numpy>=1.27
treecorr
unyt
swiftsimio
psutil
numba
matplotlib
scipy
PyYAML

